requirements/base.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
pillow==9.4.0
mysqlclient==2.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
